Allow Set-Cookie on websocket handshake response
android :: okhttp3 provides a method for inserting a cookieJar to
handle cookies implicitly.  Impltemnted a cookieJar wrapper for the
ForwardingCookieHandler and passed that to okhttp3 so that cookies
from the websocket response would be available to all other applications
for this user on this device.

iOS :: the socket rocket implementation checks incomming headers to
confirm that the websocket upgrade request results in a correct
response from the service, but fields such as "Set-Cookie" were
silently dropped.  This commit ingests the "Set-Cookie" fields from
the response header and injects them into the sharedCookieStore so
that they will be implicitly returned with any following request
to the correct domain from this user / device.

Signed-off-by: Evan J Brunner <ej3@appitto.me>